+++ utils/jera_project_setup/__brick__/lib/features/force_update/force_update_view_model.dart
@@ -1,0 +1,27 @@
+import 'dart:io';
+
+import 'force_update_view.dart';
+
+class ForceUpdateViewModel extends ForceUpdateViewModelProtocol {
+  final bool isRequired;
+
+  // Constructor
+  ForceUpdateViewModel({
+    required this.isRequired,
+  });
+
+  @override
+  bool get showIgnoreButton => !isRequired;
+
+  @override
+  String get storeURL {
+    if (Platform.isAndroid) return 'TODO';
+    return 'TODO';
+  }
+
+  @override
+  String get description {
+    if (isRequired) return 'A atualização é obrigatória';
+    return 'Atualize o app e aproveite as novas funcionalidades';
+  }
+}