+++ utils/jera_project_setup/__brick__/{{project_name.snakeCase()}}/android/app/src/main/kotlin/com/example/mason_test/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.mason_test
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}