+++ utils/jera_project_setup/CHANGELOG.md
@@ -1,3 +1,6 @@
-# 0.1.0+1
+# 0.2.3+5
+- Melhorias na funcionalidade de Force Update e adição do Fastlane
 
-- Into this version, start app project setup with flutter localization implemented, mobile router for navigation implemented, application themes and colors implemented and every api setup with dio and pretty_dio_logger.
+# 0.2.0+4
+
+- Refatorando a brick para também ser possível fazer setup de módulos Flutter e adicionando algumas funcionalidades como o force update, go router e service locator.